chore: tests & handler refactor
