--- provisioning/internal/api/export_test.go
@@ -1,4 +1,0 @@
-package api
-
-// authRequest is not exported, expose it for tests
-var AuthRequest = authRequest
